requirements.txt
# Core framework
langchain

# Community integrations (tools, loaders, vector stores)
langchain-community

# Model adapters
langchain-openai
langchain-anthropic
langchain-google-genai     # LangChain adapter for Google Gemini
langchain-deepseek

# Provider SDKs required by adapters
google-generativeai        # Official Gemini SDK required by langchain-google-genai

# Tools used by your agent
duckduckgo-search          # DuckDuckGo search wrapper
ddgs                       # Backend required by DuckDuckGoSearchRun
wikipedia                  # Wikipedia client used by WikipediaAPIWrapper

# Utilities
python-dotenv              # Load API keys from .env
pydantic                   # Typed schemas and validation
